requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
asn1crypto==1.3.0
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
clyent==1.2.2
colorama==0.4.3
conda==4.9.2
conda-build==3.18.11
conda-package-handling==1.6.0
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
glob2==0.7
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
lazy-object-proxy==1.4.3
libarchive-c==2.8
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
olefile==0.46
openpyxl==3.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
py4j==0.10.9
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pyOpenSSL==19.1.0
pyrsistent==0.15.7
PySocks==1.7.1
pyspark==3.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
requests==2.25.1
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.26.2
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xmltodict==0.12.0
zipp==2.2.0
